Conflicts:
	qtpyvcp/widgets/display_widgets/vtk_backplot/tool_actor.py
	qtpyvcp/widgets/display_widgets/vtk_backplot/vtk_backplot.py

+++ qtpyvcp/actions/machine_actions.py
@@ -401,7 +401,7 @@ def _feed_override_bindOk(value=100, widget=None):
         except AttributeError:
             widget.setValue(100)
             STATUS.feedrate.onValueChanged(
-                lambda v: widget.setValue(v * 100))
+                lambda v: widget.setValue(int(v * 100)))
 
         feed_override.set(100)
